🐞 fix: fix error test for register new user 🐞 fix: fix redirect user no team

+++ app/Actions/Fortify/CreateNewUser.php
@@ -45,8 +45,11 @@ class CreateNewUser implements CreatesNewUsers
 
         if($input['email'] && $registeruser){
             $role = 'admin';
-            $newInvitation = RoleInvitation::where('email', $input['email'])->first();
+            $team = 0;
             $newTeamMember = Jetstream::findUserByEmailOrFail($input['email']);
+
+            // ADMIN INVITATION
+            $newInvitation = RoleInvitation::where('email', $input['email'])->first();
             if($newInvitation){
                 $role = $newInvitation->role ? $newInvitation->role->name : 'admin';
                 $roleUser = RoleUser::where('user_id', $newTeamMember->id)->count();
@@ -60,19 +63,24 @@ class CreateNewUser implements CreatesNewUsers
                 }
                 $newInvitation->delete();
             }
+            // IF ADMIN INVITATION
+            // ELSE USER TEAM INVITATION
             if($newInvitation && $newInvitation->team_id){
                 $team = Team::find($newInvitation->team_id);
             }else{
                 $newInvitation = TeamInvitation::where('email', $input['email'])->first();
-                $team = Team::find($newInvitation->team_id);
-                $newTeamMember->update(['reff_team_id'=>$newInvitation->team_id]);
-                $role = $newInvitation->role;
-                $newInvitation->delete();
-
+                if($newInvitation){
+                    $team = Team::find($newInvitation->team_id);
+                    $newTeamMember->update(['reff_team_id'=>$newInvitation->team_id]);
+                    $role = $newInvitation->role;
+                    $newInvitation->delete();
+                }
             }
-            $team->users()->attach(
-                $newTeamMember, ['role' => $role]
-            );
+            if($newInvitation && $team){
+                $team->users()->attach(
+                    $newTeamMember, ['role' => $role]
+                );
+            }
         }
 
         return $registeruser;
